Revise test packages
