+++ library/src/androidTest/mock/com/mobilesolutionworks/android/app/test/basic/TestWorksDialogFragment.java
@@ -1,4 +1,4 @@
-package com.mobilesolutionworks.android.app.test.configuration;
+package com.mobilesolutionworks.android.app.test.basic;
 
 import android.app.Dialog;
 import android.content.DialogInterface;
@@ -9,7 +9,7 @@ import android.support.v7.app.AlertDialog;
 
 import com.mobilesolutionworks.android.app.WorksDialogFragment;
 import com.mobilesolutionworks.android.app.controller.WorksControllerManager;
-import com.mobilesolutionworks.android.app.test.works.GetController;
+import com.mobilesolutionworks.android.app.test.GetController;
 
 /**
  * Created by yunarta on 9/3/17.
