+++ library/src/androidTest/mock/com/mobilesolutionworks/android/app/test/nested/EmptyWorksFragment.java
@@ -1,25 +1,28 @@
-package com.mobilesolutionworks.android.app.test.works;
+package com.mobilesolutionworks.android.app.test.nested;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.TextView;
 
+import com.mobilesolutionworks.android.app.WorksFragment;
 import com.mobilesolutionworks.android.app.controller.WorksController;
 import com.mobilesolutionworks.android.app.controller.WorksControllerManager;
+import com.mobilesolutionworks.android.app.test.GetController;
 import com.mobilesolutionworks.android.app.test.R;
 
 /**
- * Created by yunarta on 12/3/17.
+ * Created by yunarta on 9/3/17.
  */
 
-public class SchedulerTestFragment extends StateWorksFragment {
+public class EmptyWorksFragment extends WorksFragment implements GetController<WorksController> {
 
     private WorksController mController;
 
     @Override
-    public void onCreate(Bundle savedInstanceState) {
+    public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         mController = getControllerManager().initController(0, null, new WorksControllerManager.ControllerCallbacks<WorksController>() {
             @Override
@@ -36,17 +39,15 @@ public class SchedulerTestFragment extends StateWorksFragment {
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.scheduler_test_fragment, null, false);
+        return inflater.inflate(R.layout.empty_fragment, container, false);
     }
 
     @Override
-    public void onResume() {
-        super.onResume();
-    }
+    public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
 
-    @Override
-    public void onPause() {
-        super.onPause();
+        TextView id = (TextView) view.findViewById(com.mobilesolutionworks.android.app.test.R.id.textView);
+        id.setText("empty");
     }
 
     @Override
